Adding fused rotation operator
